Added pluralize to minijinja-contrib
