+++ minijinja-contrib/Cargo.toml
@@ -16,3 +16,6 @@ rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "doc-header.html"]
 
 [dependencies]
 minijinja = { version = "0.31.1", path = "../minijinja", default-features = false }
+
+[dev-dependencies]
+minijinja = { version = "0.31.1", path = "../minijinja", features = ["source"] }
