+++ minijinja-contrib/src/lib.rs
@@ -1,0 +1,6 @@
+//! MiniJinja-Contrib is a utility crate for [MiniJinja](https://github.com/mitsuhiko/minijinja)
+//! that adds support for certain utilities that are too specific for the MiniJinja core.  This is
+//! usually because they provide functionality that Jinja2 itself does not have.
+
+/// Utility filters.
+pub mod filters;
